Redirecionamento do cadastro adicionado

+++ Formulário/database/available_tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user (
-    cpf INTEGER NOT NULL,
+    cpf BIGINT NOT NULL,
     nome VARCHAR(45) NOT NULL,
     email VARCHAR(45) NOT NULL,
     senha VARCHAR(20) NOT NULL,
